moved clear balance test
